+++ util/moremcp.ts
@@ -22,7 +22,12 @@ export type CallToolRequest = z.infer<typeof types.CallToolRequestSchema>
 
 export type ToolInputSchema = types.ListToolsResult["tools"][0]["inputSchema"]
 
-// eslint-disable-next-line typescript/no-empty-object-type
-export function toInputSchema<T extends z.ZodObject<{}, "strip", z.ZodTypeAny, {}, {}>>(o: T): ToolInputSchema {
+export type ToolOutputSchema = types.ListToolsResult["tools"][0]["outputSchema"]
+
+export function toInputSchema<T extends z.ZodObject<z.ZodRawShape>>(o: T): ToolInputSchema {
 	return zodToJsonSchema.zodToJsonSchema(o) as ToolInputSchema
+}
+
+export function toOutputSchema<T extends z.ZodObject<z.ZodRawShape>>(o: T): ToolOutputSchema {
+	return zodToJsonSchema.zodToJsonSchema(o) as ToolOutputSchema
 }